requirements.txt
notion-client
pandas
openpyxl
fastapi uvicorn google-generativeai
python-dotenv